Added insert workout

+++ app/src/main/java/com/mobdeves14/cadaolimyongco/Workout.kt
@@ -7,12 +7,12 @@ import androidx.room.PrimaryKey
 data class Workout (
     @PrimaryKey(autoGenerate = true) val id: Int,
     var distance: Double,
-    var duration: Int,
+    var duration: Long, // in seconds
     var avgSpeed: Double,
     var date: String,
     var monthDay: String,
     var weekDay: String,
-    var calories: Int,
+    var calories: Double,
 ) {
-    constructor(distance: Double, duration: Int, avgSpeed: Double, date: String, monthDay: String, weekDay: String, calories: Int):this(0, distance, duration, avgSpeed, date, monthDay, weekDay, calories)
+    constructor(distance: Double, duration: Long, avgSpeed: Double, date: String, monthDay: String, weekDay: String, calories: Double):this(0, distance, duration, avgSpeed, date, monthDay, weekDay, calories)
 }
